+++ src/com/github/yuthura/bianca/types/FloatType.java
@@ -5,6 +5,10 @@ import java.sql.*;
 import com.github.yuthura.bianca.*;
 
 public class FloatType extends AbstractType<Float> {
+	public FloatType() {
+		super(Float.class);
+	}
+
 	@Override
 	public int getRawSQLType() {
 		return Types.FLOAT;