+++ src/com/github/yuthura/bianca/types/LongType.java
@@ -5,6 +5,10 @@ import java.sql.*;
 import com.github.yuthura.bianca.*;
 
 public class LongType extends AbstractType<Long> {
+	public LongType() {
+		super(Long.class);
+	}
+
 	@Override
 	public int getRawSQLType() {
 		return Types.BIGINT;