+++ changelog.md
@@ -19,3 +19,4 @@
 
 - Refactored implementation of queries; impl.* package is no longer available, nor is support for OO building of queries (e.g. `select.setWhere(new Where())`)
 - Added `now` function
+- Insert queries now return the generated keys of the type specified (or default to integers)